created navigatior for daily and hourly pages
