+++ components/DailyWeather.js
@@ -1,0 +1,10 @@
+import React from "react"
+import { Text } from "react-native"
+
+import colors from "../utils/colors"
+
+export default function DailyWeather() {
+  return (
+    <Text>Daily Weather Screen</Text>
+  )
+}
