+++ components/HourlyWeather.js
@@ -1,0 +1,10 @@
+import React from "react"
+import { Text } from "react-native"
+
+import colors from "../utils/colors"
+
+export default function HourlyWeather() {
+  return (
+    <Text>Hourly Weather Screen</Text>
+  )
+}
