+++ components/WeatherDetails.js
@@ -8,7 +8,7 @@ import {
   faCloudShowersHeavy,
 } from "@fortawesome/free-solid-svg-icons"
 
-import colors from "../utils/index"
+import colors from "../utils/colors"
 
 export default function WeatherDetails({
   currentWeather,
