+++ components/WeatherMainInfo.js
@@ -1,7 +1,7 @@
 import React from "react"
 import { Text, View, SafeAreaView, StyleSheet, Image } from "react-native"
 
-import colors from "../utils/index"
+import colors from "../utils/colors"
 
 export default function WeatherMainInfo({ currentWeather }) {
   const tempCelsius = Math.floor(currentWeather.main.temp)
